Reorder some outputs Output subject name of the subject we are listing
Output subject name of the subject we are listing, and report rows
ahead of the subject being listed.

Was previously outputting the name of the previous subject, which
was confusing. By moving after the test for blank subject, we also
avoid the confusing output of a blank subject name the first time
through the loop.

Rows were reported early in the loop, meaning that the first row
of the next subject was listed with the rows from the current
subject.