Fix range check in get_by_index
get_by_index tries to check that an index is valid by using 'in',
but should be checking that it is in the range 0 to final index of
array.

If this happens, it returns self.items.

I'm not sure if this is useful and I like things to fail noisily,
so commenting out the original code and replacing with asserts
until I understand these conditions better.