Handle case where same Task appears more than once while assembling a record
I think that this can never happen with 'raw' Zooniverse workflows,
but I shall soon be adding callbacks that can insert tasks into
the workflow (e.g. where lines in a text box should be treated as
separate records, or where comma-separated parts of a line should
be treated as separate fields).