Add code to handle dropdowns
As the inline comment says, this is hacky: we just convert the
dropdowns into strings and align them as normal. This is likely to
work well enough, but it is not ideal -- for example, what happens
when 2 people pick one value and a 3rd person picks a different one?
I probably want the first value, or maybe both values -- what I'll
get is probably either the first value or some single amalgam of
both values.