Expand out multichoices in the task queue
Relies upon our being able to have multiple tasks of the same
type within a record, which the previous commit gave us.

We put a suffix (_mc, for multichoice) on the output tasks in
the queue so that we can distinguish between the original task
(with the list of all selected choices) and the new tasks (each
of which maps to one choice).