Add handling for standard minutes on attendance page
This builds on previous commits. The multichoice is expanded into
tasks-per-choice. A callback is added to break each of those
choices into one task per field and to add a dummy task which will
be used as a record boundary. The task lists are updated
accordingly.

In task names, the suffix _sm_f indicates a (f)ield added by the
(s)tandard_(m)inute callback. sm_r indicates a (r)ecord added by
the same callback.