Add capacity to extend the annotation queue with callbacks
Callbacks take the current 'ann' (a dict mapping task identifier
to a value -- usually a Zooniverse task number mapping to the
value that a volunteer gave in a classification) and return an
array of tasks to add to the queue (which will end up being fields
if they are added).

This mechanism can also be used to add record boundaries by
emitting a task which is included in the 'close' and 'add' actions.
See next commit for an example.