Cope with cases where beginnings of strings do not match at all
This may be worth checking with Mark, but I believe that what I
am doing here is handling the case where we have stopped before
traversing the whole matrix because the unexplored matrix consists
entirely of 0s. I think that this corresponds to 'none of the
remaining characters in each string actually match anything in
the other string'.

Also add a case to the __main__ condition in record_aligner_class.py.
This case fails without this change.