Improve behaviour around output directory
If it does not exist, just create it. This is always safe.

If it exists but is empty, fail unless user overrides. If user
overrides, do not delete anything just proceed -- files in output
directory should be overwritten. This is also always safe.

Deleting the directory would not be always safe. so don't do that,
force the user to do it themselves if that is what they want.